model/requirements.txt
torch
transformers
opencv-python
Pillow
pyttsx3
